After discussion with Hugo.
I realized that it does not make much sense to keep Init/ separate
as it would e.g. prevent from changing the definition of iff,
and that some stuff (plugins) depend on definitions outside of Init/
anyways.

We also think that the question of keeping plugins where they are
or moving them to the new repository must be asked more explicitely.