Rocq long-term roadmap.
This is the text as it currently appears on https://rocq-prover.org/roadmap.html.
This text was the result of discussions and meetings within the Coq core team during 2024.
The text is being added to the CEP repository in order for any community member to be able to propose changes.